% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-binarize_expr.R
\name{binary_spa_binarize}
\alias{binary_spa_binarize}
\title{Binarize expression matrix}
\usage{
binary_spa_binarize(expr_sub, threshold = 0)
}
\arguments{
\item{expr_sub}{A cells x genes expression matrix.}

\item{threshold}{Numeric cutoff (default 0).}
}
\value{
dgCMatrix (cells x genes), values 0 or 1.
}
\description{
Converts a cells x genes expression matrix into a binary 0/1 matrix.
}
